Use specialization constants for TAA flags
Generates smaller AMD ISA for 680M, but also uses up more registers and
wave64 instead of the previous wave32. Perf seems similar though.

+++ src/render/TemporalAntiAliasing.hpp
@@ -25,12 +25,14 @@
 class TemporalAntiAliasing
 {
   public:
-    enum class ColorClippingType : uint8_t
+    // NOLINTNEXTLINE(performance-enum-size) specialization constant
+    enum class ColorClippingType : uint32_t
     {
         COLOR_CLIPPING_TYPES_AND_COUNT
     };
 
-    enum class VelocitySamplingType : uint8_t
+    // NOLINTNEXTLINE(performance-enum-size) specialization constant
+    enum class VelocitySamplingType : uint32_t
     {
         VELOCITY_SAMPLING_TYPES_AND_COUNT
     };
